Move utils in their own package
